feat(litert-lm): Migrate from MediaPipe to LiteRT-LM
LiteRT-LM is the new on-device LLM inference SDK based on LiteRT Next, providing more sophisticated features specifically for large language models
(In a sense, it integrates the capabilities of ai-edge-apis and MediaPipe within one single library).

This commit replaces the MediaPipe API with the latest LiteRT-LM Kotlin API (v0.0.0-alpha05).

BREAKING CHANGE: LiteRT-LM now handles special tokens and chat templates internally.
The new API does not support pre-processing of input or post-processing of output.

As a side effect, the chat template settings UI is disabled. A follow-up refactor is required to adapt to the new API's built-in conversation management.

KNOWN ISSUE: Temperature range error, one unit test is ignored.

References:

- https://github.com/google-ai-edge/LiteRT-LM/blob/main/kotlin/README.md#litert-lm-android-api
- https://github.com/google-ai-edge/LiteRT/tree/v2.0.3?tab=readme-ov-file#litert
