+++ gradle/libs.versions.toml
@@ -15,8 +15,7 @@ kotlinxSerializationJson = "1.8.1"
 kotlinxSerializationPlugin = "2.2.0"
 onnxRuntime = "1.23.2"
 executorch = "1.0.0"
-mediapipeTasksText = "0.10.26"
-mediapipeTasksGenai = "0.10.25"
+litertlm = "0.0.0-alpha05"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -52,9 +51,9 @@ onnxruntime-android = { group = "com.microsoft.onnxruntime", name = "onnxruntime
 # The `-vulkan` build includes Vulkan backend in addition to XnnpackBackend.
 # Docs: https://docs.pytorch.org/executorch/1.0/using-executorch-android.html
 executorch-android = { group = "org.pytorch", name = "executorch-android-vulkan", version.ref = "executorch" }
-# MediaPipe (LiteRT).
-mediapipe-tasks-text = { group = "com.google.mediapipe", name = "tasks-text", version.ref = "mediapipeTasksText" }
-mediapipe-tasks-genai = { group = "com.google.mediapipe", name = "tasks-genai", version.ref = "mediapipeTasksGenai" }
+# LiteRT-LM.
+# See https://github.com/google-ai-edge/LiteRT-LM/tree/main/kotlin#litert-lm-android-api
+litertlm = { group = "com.google.ai.edge.litertlm", name = "litertlm", version.ref = "litertlm" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
