--- mediapipe/src/androidTest/assets/README.md
@@ -1,2 +1,0 @@
-The test model file copied from:
-https://github.com/google-ai-edge/LiteRT-LM/tree/v0.7.0/runtime/testdata
